# Conflicts:
#	mLRS/Common/bind.h
#	tools/run_make_firmwares.py

+++ mLRS/CommonRx/mlrs-rx.cpp
@@ -169,8 +169,8 @@ void init_hw(void)
 // SX12xx
 //-------------------------------------------------------
 
-volatile uint16_t irq_status;
-volatile uint16_t irq2_status;
+volatile uint32_t irq_status;
+volatile uint32_t irq2_status;
 
 IRQHANDLER(
 void SX_DIO_EXTI_IRQHandler(void)
